pyproject.toml
[tool.poetry]
name = "bas-air-unit-waypoints"
version = "0.1.0"
description = "Data management for waypoints and routes within the British Antarctic Survey Air Unit"
authors = ["Felix Fennell <felnne@bas.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.2"
ulid-py = "^1.1.0"
Shapely = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[tool.black]
line-length = 120
target-version = ['py36']
include = '\.pyi?$'
exclude = ''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
